add location manager

+++ app/src/main/java/ru/test/weather/data/network/IApiService.kt
@@ -8,7 +8,7 @@ import ru.test.weather.data.models.weather.response.WeatherByCoordinatesResponse
 interface IApiService {
 
     @GET("weather")
-    fun getWeatherByCoordinates(@Query("lat") latitude: Float, @Query("lon") longitude: Float,
+    fun getWeatherByCoordinates(@Query("lat") latitude: Double, @Query("lon") longitude: Double,
                                 @Query("units") unitsFormat: String,
                                 @Query("lang") language: String): Single<WeatherByCoordinatesResponse>
 }
