+++ app/src/main/java/ru/test/weather/domain/models/location/WeatherPoint.kt
@@ -1,4 +1,4 @@
 package ru.test.weather.domain.models.location
 
-data class WeatherPoint(val latitude: Float, val longitude: Float) {
+data class WeatherPoint(val latitude: Double, val longitude: Double) {
 }
